+++ azimuth/modules/base_classes/aggregation_module.py
@@ -3,12 +3,13 @@
 # in the root directory of this source tree.
 import time
 from abc import ABC
-from typing import List, Optional
+from typing import List, Optional, cast
 
 from datasets import Dataset
 
 from azimuth.modules.base_classes import ConfigScope, ExpirableMixin, Module
-from azimuth.types import DatasetSplitName, ModuleOptions, ModuleResponse
+from azimuth.types import DatasetColumn, DatasetSplitName, ModuleOptions, ModuleResponse
+from azimuth.types.outcomes import OutcomeName
 from azimuth.utils.filtering import filter_dataset_split
 
 
@@ -33,7 +34,7 @@ class ComparisonModule(AggregationModule[ConfigScope], ABC):
 class FilterableModule(AggregationModule[ConfigScope], ExpirableMixin, ABC):
     """Filterable Module are affected by filters in mod options."""
 
-    allowed_mod_options = {"filters", "pipeline_index"}
+    allowed_mod_options = {"filters", "pipeline_index", "without_postprocessing"}
 
     def __init__(
         self,
@@ -56,3 +57,42 @@ class FilterableModule(AggregationModule[ConfigScope], ExpirableMixin, ABC):
         """
         ds = super().get_dataset_split(name)
         return filter_dataset_split(ds, filters=self.mod_options.filters, config=self.config)
+
+    def _get_predictions_from_ds(self) -> List[int]:
+        """Get predicted classes according to the module options (with or without postprocessing).
+
+        Returns: List of Predictions
+        """
+        ds = self.get_dataset_split()
+        if self.mod_options.without_postprocessing:
+            return cast(List[int], [preds[0] for preds in ds[DatasetColumn.model_predictions]])
+        else:
+            return cast(List[int], ds[DatasetColumn.postprocessed_prediction])
+
+    def _get_confidences_from_ds(self) -> List[List[float]]:
+        """Get confidences according to the module options (with or without postprocessing).
+
+        Notes: Confidences are sorted according to their values (not the class id).
+
+        Returns: List of Confidences
+        """
+        ds = self.get_dataset_split()
+        confidences = (
+            ds[DatasetColumn.model_confidences]
+            if self.mod_options.without_postprocessing
+            else ds[DatasetColumn.postprocessed_confidences]
+        )
+        return cast(List[List[float]], confidences)
+
+    def _get_outcomes_from_ds(self) -> List[OutcomeName]:
+        """Get outcomes according to the module options (with or without postprocessing).
+
+        Returns: List of Outcomes
+        """
+        ds = self.get_dataset_split()
+        outcomes = (
+            ds[DatasetColumn.model_outcome]
+            if self.mod_options.without_postprocessing
+            else ds[DatasetColumn.postprocessed_outcome]
+        )
+        return cast(List[OutcomeName], outcomes)