+++ azimuth/routers/v1/model_performance/confusion_matrix.py
@@ -38,10 +38,12 @@ def get_confusion_matrix(
     task_manager: TaskManager = Depends(get_task_manager),
     dataset_split_manager: DatasetSplitManager = Depends(get_dataset_split_manager),
     pipeline_index: int = Depends(require_pipeline_index),
+    without_postprocessing: bool = False,
 ) -> ConfusionMatrixResponse:
     mod_options = ModuleOptions(
         filters=named_filters.to_dataset_filters(dataset_split_manager.get_class_names()),
         pipeline_index=pipeline_index,
+        without_postprocessing=without_postprocessing,
     )
 
     task_result: ConfusionMatrixResponse = get_standard_task_result(