+++ azimuth/routers/v1/model_performance/metrics.py
@@ -46,10 +46,12 @@ def get_metrics(
     task_manager: TaskManager = Depends(get_task_manager),
     dataset_split_manager: DatasetSplitManager = Depends(get_dataset_split_manager),
     pipeline_index: int = Depends(require_pipeline_index),
+    without_postprocessing: bool = False,
 ) -> MetricsAPIResponse:
     mod_options = ModuleOptions(
         filters=named_filters.to_dataset_filters(dataset_split_manager.get_class_names()),
         pipeline_index=pipeline_index,
+        without_postprocessing=without_postprocessing,
     )
 
     module_response: List[MetricsModuleResponse] = get_standard_task_result(