+++ azimuth/routers/v1/model_performance/outcome_count.py
@@ -78,10 +78,12 @@ def get_outcome_count_per_filter(
     task_manager: TaskManager = Depends(get_task_manager),
     dataset_split_manager: DatasetSplitManager = Depends(get_dataset_split_manager),
     pipeline_index: int = Depends(require_pipeline_index),
+    without_postprocessing: bool = False,
 ) -> OutcomeCountPerFilterResponse:
     mod_options = ModuleOptions(
         filters=named_filters.to_dataset_filters(dataset_split_manager.get_class_names()),
         pipeline_index=pipeline_index,
+        without_postprocessing=without_postprocessing,
     )
 
     task_result: OutcomeCountPerFilterResponse = get_standard_task_result(