+++ tests/test_modules/test_word_analysis/test_top_words.py
@@ -4,15 +4,10 @@
 
 import numpy as np
 
-from azimuth.modules.model_contracts import FileBasedTextClassificationModule
 from azimuth.modules.word_analysis.top_words import TopWordsModule
-from azimuth.types import (
-    DatasetFilters,
-    DatasetSplitName,
-    ModuleOptions,
-    SupportedMethod,
-)
+from azimuth.types import DatasetFilters, DatasetSplitName, ModuleOptions
 from azimuth.types.word_analysis import TokensToWordsResponse
+from tests.utils import save_predictions
 
 ALL_WORDS = [
     "hello",
@@ -77,19 +72,7 @@ def fake_no_saliency(indices):
 
 
 def test_top_words_without_saliency(monkeypatch, file_text_config_top1, dask_client):
-
-    # Get predictions
-    pred_mod = FileBasedTextClassificationModule(
-        DatasetSplitName.eval,
-        file_text_config_top1,
-        mod_options=ModuleOptions(
-            pipeline_index=0, model_contract_method_name=SupportedMethod.Predictions
-        ),
-    )
-
-    res = pred_mod.compute_on_dataset_split()
-
-    pred_mod.save_result(res, pred_mod.get_dataset_split_manager(DatasetSplitName.eval))
+    save_predictions(file_text_config_top1)
 
     mod = TopWordsModule(
         dataset_split_name=DatasetSplitName.eval,