+++ tests/test_utils/test_dataset_filtering.py
@@ -1,7 +1,6 @@
 # Copyright ServiceNow, Inc. 2021 – 2022
 # This source code is licensed under the Apache 2.0 license found in the LICENSE file
 # in the root directory of this source tree.
-import math
 
 import pytest
 
@@ -11,7 +10,7 @@ from azimuth.types.tag import ALL_SMART_TAGS, DataAction, SmartTag
 from azimuth.utils.filtering import filter_dataset_split
 
 
-def test_dataset_filtering(text_dm_with_tags, simple_text_config, simple_table_key):
+def test_dataset_filtering(text_dm_with_tags, simple_table_key):
     ds = text_dm_with_tags.get_dataset_split(simple_table_key)
 
     ds_len = {}
@@ -22,27 +21,27 @@ def test_dataset_filtering(text_dm_with_tags, simple_text_config, simple_table_k
         return len(ds_filtered)
 
     assert filtered_len(DatasetFilters()) == len(ds)
-    assert filtered_len(DatasetFilters(confidence_min=0.5)) == len(ds) // 2
-    assert filtered_len(DatasetFilters(confidence_max=0.5)) == len(ds) // 2 + 1
+    assert filtered_len(DatasetFilters(confidence_min=0.5)) == 0.9 * len(ds) // 2
+    assert filtered_len(DatasetFilters(confidence_max=0.5)) == len(ds) - 0.9 * len(ds) // 2
     assert filtered_len(DatasetFilters(labels=[0])) == 22
     assert filtered_len(DatasetFilters(data_actions=[DataAction.relabel])) == 1
-    assert filtered_len(DatasetFilters(outcomes=[OutcomeName.IncorrectAndRejected])) == 22
+    assert filtered_len(DatasetFilters(outcomes=[OutcomeName.IncorrectAndRejected])) == 33
     assert filtered_len(DatasetFilters(smart_tags=[SmartTag.short])) == 1
     assert filtered_len(DatasetFilters(utterance="some")) == 2
-    assert filtered_len(DatasetFilters(predictions=[1])) == 10
+    assert filtered_len(DatasetFilters(predictions=[1])) == 5
 
     # We can filter by combinations of filter
     combination_len = filtered_len(DatasetFilters(labels=[0], predictions=[1]))
-    assert combination_len == 6
+    assert combination_len == 3
     assert combination_len < min(ds_len[((), (1,))], ds_len[((0,), ())])
 
 
-def test_dataset_filtering_errors(text_dm_with_tags, simple_text_config, simple_table_key):
+def test_dataset_filtering_errors(text_dm_with_tags, simple_table_key):
     ds = text_dm_with_tags.get_dataset_split(simple_table_key)
     ds = ds.rename_column(DatasetColumn.postprocessed_prediction, "col1")
     ds = ds.rename_column("label", "col2")
     ds = ds.rename_column(text_dm_with_tags.config.columns.text_input, "col3")
-    ds = ds.rename_column(DatasetColumn.outcome, "col4")
+    ds = ds.rename_column(DatasetColumn.postprocessed_outcome, "col4")
     ds = ds.rename_column(DatasetColumn.postprocessed_confidences, "col5")
 
     with pytest.raises(ValueError) as e:
@@ -61,15 +60,15 @@ def test_dataset_filtering_errors(text_dm_with_tags, simple_text_config, simple_
         _ = filter_dataset_split(ds, DatasetFilters(labels=[1]), config=text_dm_with_tags.config)
     assert "label" in str(e.value)
 
-    with pytest.raises(KeyError) as e:
+    with pytest.raises(ValueError) as e:
         _ = filter_dataset_split(
             ds,
             DatasetFilters(outcomes=[OutcomeName.IncorrectAndRejected]),
             config=text_dm_with_tags.config,
         )
-    assert DatasetColumn.outcome in str(e.value)
+    assert DatasetColumn.postprocessed_outcome in str(e.value)
 
-    with pytest.raises(KeyError) as e:
+    with pytest.raises(ValueError) as e:
         _ = filter_dataset_split(
             ds, DatasetFilters(confidence_max=0.5), config=text_dm_with_tags.config
         )
@@ -128,7 +127,7 @@ def test_dataset_filtering_confidence(text_dm_with_tags, simple_table_key):
     ds_filtered = filter_dataset_split(
         ds, DatasetFilters(confidence_min=0.4, confidence_max=0.6), config=text_dm_with_tags.config
     )
-    assert len(ds_filtered) == math.ceil(len(ds) / 5)
+    assert len(ds_filtered) == 10
     assert (
         len(
             ds_filtered.filter(
@@ -147,6 +146,22 @@ def test_dataset_filtering_mutually_exclusive(text_dm_with_tags, simple_table_ke
         config=text_dm_with_tags.config,
     )
     assert len(ds_filtered) == 0
+
+
+def test_dataset_filtering_without_postprocessing(text_dm_with_tags, simple_table_key):
+    ds = text_dm_with_tags.get_dataset_split(simple_table_key)
+    ds_filtered_with_postprocessing = filter_dataset_split(
+        ds,
+        DatasetFilters(predictions=[0]),
+        config=text_dm_with_tags.config,
+    )
+    ds_filtered_without_postprocessing = filter_dataset_split(
+        ds,
+        DatasetFilters(predictions=[0]),
+        config=text_dm_with_tags.config,
+        without_postprocessing=True,
+    )
+    assert len(ds_filtered_with_postprocessing) != len(ds_filtered_without_postprocessing)
 
 
 if __name__ == "__main__":